delete redundant files
