+++ fsae-vehicle-fw/src/vehicle/devices/shockTravel.cpp
@@ -2,7 +2,7 @@
 
 #include "shockTravel.h"
 #include "utils/utils.h"
-#include "vehicle/telemetry.h"
+#include "vehicle/comms/telemetry.h"
 #include <arduino_freertos.h>
 #include <cmath>
